requirements.txt
requests==2.31.0
fastapi==0.104.1
mangum==0.17.0
pydantic==2.5.1
aws-lambda-powertools==2.28.0
jmespath==1.0.1
pytest==7.4.3
python-dotenv==1.0.0
redis==4.6.0
supabase==2.1.0
sqlalchemy==2.0.23
uvicorn==0.24.0
boto3==1.29.6
botocore==1.32.6
loguru==0.7.2